fix(web): close tenant-boundary seams in dashboard server actions
- Centralize org-scoped Prisma in dashboard-org-scoped-prisma helpers
- Align remediation/review mutations with active org + scoped where clauses
- Move operator console queries out of actions.ts; export types from shared module
- Wrap scan enqueue via helpers; tighten crawl config updates with org filter
- Add unit tests for scoped query shapes; update auto-scan action test mocks
