+++ apps/web/src/app/(dashboard)/remediation/[suggestionId]/actions.ts
@@ -2,78 +2,35 @@
 
 import { redirect } from "next/navigation";
 import { requireSession } from "@/lib/session";
-import { prisma } from "@/lib/db";
 import { hasPermission } from "@aros/config";
-import { getEntitlementState, requireOrgAccess } from "@/lib/auth-guard";
-
-async function requireSuggestionAccess(suggestionId: string, userId: string) {
-  if (!suggestionId) {
-    redirect("/remediation?error=missing_id");
-  }
-
-  // First get the suggestion and determine the organization
-  const suggestion = await prisma.remediationSuggestion.findUnique({
-    where: { id: suggestionId },
-    select: {
-      id: true,
-      status: true,
-      recipeId: true,
-      finding: {
-        select: {
-          site: {
-            select: {
-              workspace: {
-                select: { organizationId: true },
-              },
-            },
-          },
-        },
-      },
-      cluster: {
-        select: {
-          site: {
-            select: {
-              workspace: {
-                select: { organizationId: true },
-              },
-            },
-          },
-        },
-      },
-    },
-  });
-
-  if (!suggestion) {
-    redirect("/remediation?error=not_found");
-  }
-
-  const organizationId =
-    suggestion.finding?.site.workspace.organizationId ??
-    suggestion.cluster?.site.workspace.organizationId;
-
-  if (!organizationId) {
-    redirect("/remediation?error=invalid_suggestion");
-  }
-
-  // Use centralized auth guard
-  const ctx = await requireOrgAccess(organizationId, "suggestion:approve", {
-    requirePaid: true,
-  });
-
-  return {
-    suggestion,
-    role: ctx.role,
-    entitlement: ctx.entitlement,
-  };
-}
+import { requireOrgAccess } from "@/lib/auth-guard";
+import { getRoutePlatformTruth } from "@/lib/platform-truth-cache";
+import { resolveDashboardOrgMembership } from "@/lib/route-data-boundary";
+import {
+  loadRemediationSuggestionForOrg,
+  updateRemediationSuggestionApproved,
+  updateRemediationSuggestionExported,
+  updateRemediationSuggestionRejected,
+} from "@/lib/dashboard-org-scoped-prisma";
 
 export async function approveSuggestionAction(formData: FormData) {
   const user = await requireSession();
   const suggestionId = (formData.get("suggestionId") as string) || "";
 
-  const { suggestion, role, entitlement } = await requireSuggestionAccess(
-    suggestionId,
-    user.id,
+  if (!suggestionId) {
+    redirect("/remediation?error=missing_id");
+  }
+
+  const platformTruth = await getRoutePlatformTruth();
+  const orgRes = await resolveDashboardOrgMembership(user.id, platformTruth);
+  if (orgRes.kind !== "ok") {
+    redirect("/remediation?error=not_found");
+  }
+
+  const { role, entitlement } = await requireOrgAccess(
+    orgRes.organizationId,
+    "suggestion:approve",
+    { requirePaid: true },
   );
 
   if (!entitlement.hasPaidAccess) {
@@ -84,26 +41,24 @@ export async function approveSuggestionAction(formData: FormData) {
     redirect(`/remediation/${suggestionId}?error=forbidden`);
   }
 
+  const suggestion = await loadRemediationSuggestionForOrg(
+    suggestionId,
+    orgRes.organizationId,
+  );
+  if (!suggestion) {
+    redirect("/remediation?error=not_found");
+  }
+
   if (suggestion.status !== "DRAFT" && suggestion.status !== "VALIDATED") {
     redirect(`/remediation/${suggestionId}?error=invalid_transition`);
   }
 
-  try {
-    await prisma.remediationSuggestion.update({
-      where: { id: suggestionId },
-      data: {
-        status: "APPROVED",
-        appliedBy: user.id,
-        appliedAt: new Date(),
-      },
-    });
-    if (suggestion.recipeId) {
-      await prisma.remediationRecipe.update({
-        where: { id: suggestion.recipeId },
-        data: { successCount: { increment: 1 } },
-      });
-    }
-  } catch {
+  const outcome = await updateRemediationSuggestionApproved(
+    suggestionId,
+    orgRes.organizationId,
+    user.id,
+  );
+  if (!outcome.ok) {
     redirect(`/remediation/${suggestionId}?error=update_failed`);
   }
 
@@ -114,9 +69,20 @@ export async function rejectSuggestionAction(formData: FormData) {
   const user = await requireSession();
   const suggestionId = (formData.get("suggestionId") as string) || "";
 
-  const { suggestion, role, entitlement } = await requireSuggestionAccess(
-    suggestionId,
-    user.id,
+  if (!suggestionId) {
+    redirect("/remediation?error=missing_id");
+  }
+
+  const platformTruth = await getRoutePlatformTruth();
+  const orgRes = await resolveDashboardOrgMembership(user.id, platformTruth);
+  if (orgRes.kind !== "ok") {
+    redirect("/remediation?error=not_found");
+  }
+
+  const { role, entitlement } = await requireOrgAccess(
+    orgRes.organizationId,
+    "suggestion:approve",
+    { requirePaid: true },
   );
 
   if (!entitlement.hasPaidAccess) {
@@ -127,22 +93,23 @@ export async function rejectSuggestionAction(formData: FormData) {
     redirect(`/remediation/${suggestionId}?error=forbidden`);
   }
 
+  const suggestion = await loadRemediationSuggestionForOrg(
+    suggestionId,
+    orgRes.organizationId,
+  );
+  if (!suggestion) {
+    redirect("/remediation?error=not_found");
+  }
+
   if (suggestion.status === "REJECTED" || suggestion.status === "APPLIED") {
     redirect(`/remediation/${suggestionId}?error=invalid_transition`);
   }
 
-  try {
-    await prisma.remediationSuggestion.update({
-      where: { id: suggestionId },
-      data: { status: "REJECTED" },
-    });
-    if (suggestion.recipeId) {
-      await prisma.remediationRecipe.update({
-        where: { id: suggestion.recipeId },
-        data: { rejectionCount: { increment: 1 } },
-      });
-    }
-  } catch {
+  const outcome = await updateRemediationSuggestionRejected(
+    suggestionId,
+    orgRes.organizationId,
+  );
+  if (!outcome.ok) {
     redirect(`/remediation/${suggestionId}?error=update_failed`);
   }
 
@@ -153,9 +120,20 @@ export async function exportSnippetAction(formData: FormData) {
   const user = await requireSession();
   const suggestionId = (formData.get("suggestionId") as string) || "";
 
-  const { suggestion, role, entitlement } = await requireSuggestionAccess(
-    suggestionId,
-    user.id,
+  if (!suggestionId) {
+    redirect("/remediation?error=missing_id");
+  }
+
+  const platformTruth = await getRoutePlatformTruth();
+  const orgRes = await resolveDashboardOrgMembership(user.id, platformTruth);
+  if (orgRes.kind !== "ok") {
+    redirect("/remediation?error=not_found");
+  }
+
+  const { role, entitlement } = await requireOrgAccess(
+    orgRes.organizationId,
+    "suggestion:export",
+    { requirePaid: true },
   );
 
   if (!entitlement.hasPaidAccess) {
@@ -166,16 +144,23 @@ export async function exportSnippetAction(formData: FormData) {
     redirect(`/remediation/${suggestionId}?error=forbidden`);
   }
 
+  const suggestion = await loadRemediationSuggestionForOrg(
+    suggestionId,
+    orgRes.organizationId,
+  );
+  if (!suggestion) {
+    redirect("/remediation?error=not_found");
+  }
+
   if (suggestion.status !== "APPROVED" && suggestion.status !== "VALIDATED") {
     redirect(`/remediation/${suggestionId}?error=invalid_transition`);
   }
 
-  try {
-    await prisma.remediationSuggestion.update({
-      where: { id: suggestionId },
-      data: { status: "EXPORTED" },
-    });
-  } catch {
+  const outcome = await updateRemediationSuggestionExported(
+    suggestionId,
+    orgRes.organizationId,
+  );
+  if (!outcome.ok) {
     redirect(`/remediation/${suggestionId}?error=update_failed`);
   }
